+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - 1.2.4
     - Refactor neural network parameter updates
+    - Allow set null logger
 
 - 1.2.3
     - Fix Multiclass layer cross entropy gradient optimization